Abstract out runCommand function
